# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "backend",
    "description": "",
    "contact": {
      "name": "David Hewitt",
      "email": "davidmhewitt@gmail.com"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/dashboard/create_stripe_account": {
      "post": {
        "tags": [
          "dashboard::create_stripe_account"
        ],
        "operationId": "create",
        "responses": {
          "200": {
            "description": "Stripe account created"
          },
          "500": {
            "description": "Error occurred while creating stripe account"
          }
        }
      }
    },
    "/dashboard/enable_app_payments/{app_id}": {
      "post": {
        "tags": [
          "dashboard::enable_app_payments"
        ],
        "operationId": "enable_app_payments",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payments enabled for app"
          },
          "500": {
            "description": "Error occurred while enabling payments for app"
          }
        }
      }
    },
    "/dashboard/link_stripe_account": {
      "get": {
        "tags": [
          "dashboard::link_stripe_account"
        ],
        "operationId": "link",
        "responses": {
          "303": {
            "description": "Stripe account successfully linked, follow redirect for Stripe setup",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Stripe URL the user's browser should be redirected to"
              }
            }
          },
          "500": {
            "description": "Error occurred while linking stripe account"
          }
        }
      }
    },
...
    "/users/register/confirm": {
      "get": {
        "tags": [
          "users::confirm_registration"
        ],
        "operationId": "confirm",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "303": {
            "description": ""
          }
        }
      }
    },
...
  },
...
}